+++ art/art.py
@@ -214,17 +214,20 @@ def tprint(
         if font == "UnicodeEncodeError":
             raise UnicodeEncodeError(
                 'test', u"", 42, 43, 'test unicode-encode-error')
-        result, font = text2art(
+        result, font, decoration = text2art(
             text,
             font=font,
             decoration=decoration,
             chr_ignore=chr_ignore,
             sep=sep,
             space=space,
-            __font_return=True)
+            __detailed_return=True)
         print(result)
     except UnicodeEncodeError:
-        print(FONT_ENVIRONMENT_WARNING.format(font))
+        if decoration is not None:
+            print(FONT_OR_DECOR_ENVIRONMENT_WARNING.format(font, decoration))
+        else:
+            print(FONT_ENVIRONMENT_WARNING.format(font))
 
 
 def tsave(
@@ -476,7 +479,7 @@ def text2art(
         decoration=None,
         sep="\n",
         space=0,
-        __font_return=False):
+        __detailed_return=False):
     r"""
     Return art text (support \n).
 
@@ -492,8 +495,8 @@ def text2art(
     :type sep: str
     :param space: space between characters
     :type space: int
-    :param __font_return: flag for returning the font
-    :type __font_return: bool
+    :param __detailed_return: flag for returning the font and the decoration
+    :type __detailed_return: bool
     :return: ascii art text as str
     """
     letters = standard_dic
@@ -528,8 +531,8 @@ def text2art(
     if decoration is not None:
         [decor1, decor2] = decor(decoration, both=True)
         result = decor1 + result + decor2
-    if __font_return:
-        return (result, font)
+    if __detailed_return:
+        return (result, font, decoration)
     return result
 
 
@@ -542,7 +545,7 @@ def set_default(
         decoration=None,
         sep="\n",
         space=0,
-        __font_return=False):
+        __detailed_return=False):
     """
     Change text2art, tprint and tsave default values.
 
@@ -562,8 +565,8 @@ def set_default(
     :type sep: str
     :param space: space between characters
     :type space: int
-    :param __font_return: flag for returning the font
-    :type __font_return: bool
+    :param __detailed_return: flag for returning the font
+    :type __detailed_return: bool
     :return: None
     """
     if isinstance(font, str) is False:
@@ -582,8 +585,8 @@ def set_default(
         raise artError(SEP_TYPE_ERROR)
     if isinstance(space, int) is False:
         raise artError(SPACE_TYPE_ERROR)
-    if isinstance(__font_return, bool) is False:
-        raise artError(FONT_RETURN_TYPE_ERROR)
+    if isinstance(__detailed_return, bool) is False:
+        raise artError(DETAILED_RETURN_TYPE_ERROR)
     tprint.__defaults__ = (font, chr_ignore, decoration, sep, space)
     tsave.__defaults__ = (
         font,
@@ -600,7 +603,7 @@ def set_default(
         decoration,
         sep,
         space,
-        __font_return)
+        __detailed_return)
 
 
 def get_font_dic(font_name):