+++ art/test2.py
@@ -3062,7 +3062,9 @@ False
 >>> aprint("UnicodeEncodeError")
 [Warning] 'UnicodeEncodeError' is not printable in this environment.
 >>> tprint("test","UnicodeEncodeError")
-[Warning] 'UnicodeEncodeError' is not printable in this environment.
+[Warning] 'UnicodeEncodeError' font is not printable in this environment.
+>>> tprint("test","UnicodeEncodeError",decoration="angry1")
+[Warning] 'UnicodeEncodeError' font or 'angry1' decoration is not printable in this environment.
 >>> Data = tsave("test@34",font="antrophobia",filename="antrophobia.txt")
 Saved!
 Filename: antrophobia.txt