+++ CHANGELOG.md
@@ -6,7 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Changed
-- `__font_return` parameter added to `text2art` function
+- `__detailed_return` parameter added to `text2art` function
 - `__word2art` function modified
 - `sapce` bug fixed
 ## [6.0] - 2023-06-14